dataset stats fix
